Add directories/folders
All base folders and files that are standard with a Django project and additional files for handling webpage interaction. Planning to include empty folders meant to house the CSS and HTML files. These files will be updated to contain the same information as their respective files outside the mysite folder. The only file that will remain outside the mysite folder will be the manage.py file.