Classes (#3)
* Class creation

Classes made based off of the database. Will eventually be added to the Django framework.

* Add views and urls python files

Django code for handling login page and SQL data concerned with user log-in/sign-up.

* Update models.py

Added lines 10-12 for use with views.py

* Add settings and manage files

Django settings and manage files to be part of the Django project and connect database.

* Add directories/folders

All base folders and files that are standard with a Django project and additional files for handling webpage interaction. Planning to include empty folders meant to house the CSS and HTML files. These files will be updated to contain the same information as their respective files outside the mysite folder. The only file that will remain outside the mysite folder will be the manage.py file.

* Update settings.py

Adding relative database format/information from the previous settings.py file.

* Delete settings.py

This file has been expanded on and moved into the mysite folder.

* Delete manage.py

The new and expanded manage.py can now be found in the mysite folder.

* Delete views.py

This file now exists in the 'mysite' folder.

* Merging of front / backend

Currently am able to:
1. Run the application
2. View both the create deck and login page
3. Currently its hard to test what works without the login / signup stuff configured first.

* Connected Home HTML

Connected the home html properly with what we have so there is a centralized working location. Creating a deck works now, even without being logged in

* Can now create and view decks

- Can now create a deck via the 'library' page
- Can view what decks we have via the 'library' page
- Can also see what info we have for the deck
- Can delete the deck

* Update README.md

* Fix

* Update urls.py

* Add __init__.py

Please review this version of the __init__.py file to see if it should replace the one that already exists inside the mysite folder here in this branch.

* Update views.py

Updating views.py so it will include the early form of user authentication.

* Update urls.py

Updating URLs will assist with early user authentication features.

* Update Home.html

Updating Home.html to have the early form of user authentication with a Logout button when logged in.

* Update login.html

Introducing early concept of a message that should let the user know whether login information is incorrect or not. Correct information should result in the user being redirected to the Home.html page.

* Update views.py

* Update views.py

* Update views.py

* Update urls.py

Updating to the latest stable version of urls.py that allows navigation to the signup page.

* Update views.py

Updating to the latest stable version of views.py that allows for navigation to the signup page.

* Update Home.html

Updating to the latest stable version that allows for navigation to the signup page.

* Update sign_up.html

* Update login.html

Updated to include navigation to signup page.

* Update sign_up.html

Updated to include navigation to login page.

* Update README.md

Added newly available access to signup page.

* User functionality, deck and flashcard classes.

User functionality and fully functioning deck and flashcard classes added and incorporated with temp html / css

* Update README.md

* Delete __init__.py

An updated version of this file can now be found in the mysite folder.

* Update sign_up.html

Adding missing '</html>' to indicate the end of the document.

* Update Home.html

Ensures the signup button is also hidden for authenticated users.

* Update Library.html

Hides signup button from view for authenticated users.

* Update Settings.html

Hides the signup button from view for authenticated users.

* Update README.md

---------

Co-authored-by: rdurdin <rdurdin@my.waketech.edu>
Co-authored-by: njgarner <118210681+njgarner@users.noreply.github.com>
Co-authored-by: Raegan Durdin <139166552+rdurdin@users.noreply.github.com>