Updated Style and base.html
Created base.html, this file can be used by other template files to avoid having to copy and paste the code for the navigation header and sidebar by making the files extended form base.html.

Updated style.css to incorporate styling for the Create Deck and Create Flashcard pages

Updated create_flashcard.html to be extended from base.html to provide the navigation header and sidebar, adding styling to the page. Also made a back-button class to be able to style back buttons to be blue

Updated deck.html to be extended from base.html and provided styling for the page including a Back Button to take users back to the library page