Updated styling for flashcards
Updated Home.html to add the Deck Selector HTML code for future use, the deck selector would be used by Users to select what deck they want to study and view.Also added the "flashcard-text" class to the card text paragraph element for styling

Updated styles.css to change a few things for the .flashcard class, like more padding, rounder border, wider border, and aligning the elements inside the div to be in the middle of the card (The text). Added flashcard-text to style the text inside the flashcard. Added .flashcard-display label and select to style the labels and select elements in the flashcard-display class divs, will be used by the Deck Selector in the homepage.