Added Hint Placeholder
Updated style.css to include styling for the hint feature. hint-display is the div that the hint button and content will go to. hint-btn-on and hint-button-off are styling for the button depending on if it is showing a hint or not. hint-content is the styling for the div with the hint-content id, this is where the text will be. hint-content-on and hint-content-off are classes that will determine if a user can see a hint or not.

Updated home.html to include html code for the hint feature, also include a new script section that contains code for switching the styling class of the hint features (Hint button will switch to on or off classes depending on state it is in when pressed, also affects the hint content class)