Update Flashcard/Set limits
Updated flashcard_set_detail.html to include a card_count variable from the views.py page's "view_flashcard_set" section so users know how many flashcards are in the set. Also moved the back button to be below the set title.

Updated library.html to move the messages to the top of the screen instead of the bottom. Added a Set counter that counts the amount of sets the user has out of the max amout of sets.

Updated views.py to add "set_count" to the library_view, this will count the total number of sets a user currently has. Updated "create_set" to also have a set count and make it so that a user cannot create a new set if they already have 30 or move currently. "create_flashcard" was updated to make it so that when the submit buttton for the form is pressed, it will count the current amount of flashcards in the set that the user selected in the flashcard set field selector, then if the amount of card is 30 or more, prevents the user from creating that card to that set. "view_flashcard_set" also has the card counter.