Updated User Auth Pages/Style
Updated user_auth.css to add styling for .navigation-div class to center the logo. Changed some styling for .auth-div. Added styling for auth-heading id (The Title of the page). Changed width of the logo. Added styling for the buttons to fit with the main pages buttons.

Updated login.html to remove some <br> elements and add a "---- OR ----" divider.

Updated sign_up.html to remove <br> elements and change the heading element from h2 to h1. Added the logo, and added "btn-success" class to the submit button