Updated Create Flashcard
Updated forms.py to remove the 'flashcard_set' fields in the Flashcard form in order to prevent users from selecting a set like they could before. This is part of a slight rework to how users create cards. From now on, users will go to a Set's detail page and click the create flashcard button to go to the create_flashcard page, then the flashcard_set field in the Flashcard Model is automatically filled by the Set which you had in the details page.

Updated style.css to make the Back Buttons now change thier width to fit the content size so no more manually having to change width

Updated create_flashcard.html to display what Set you are adding the new flashcard to, and added a Back Button to go back to the Set's detail page

Updated flashcard_set_detail.html to add the flashcard_set.set_id to the create_flashcard link

Updated urls.py to add the <int:set_id> to the create_flashcard link.

Updated view.py to make it so that the flashcard form has the flashcard_set field filled out with the Set that was in the details page